requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
httpx==0.26.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
xgboost==2.0.3
joblib==1.3.2
python-multipart==0.0.6
aiofiles==23.2.1
